Added support for new spotify player
This is what I'm using in a current unpacked extension, and it seems to work fine in the United States. I did use the chrome extensions page to customize my keyboard shortcuts to media hotkeys and make the shortcuts system wide, and it still works. There is definitely some latency between the click and the command running.